implemented pr comments
