+++ src/mlpa/admin/index.html
@@ -68,12 +68,12 @@
       <div class="flex gap-3 items-end">
         <div class="flex-1">
           <label for="masterKeyInput" class="block text-sm font-medium text-gray-700 mb-1.5">
-            Master Key
+            UI Access Key
           </label>
           <input
             id="masterKeyInput"
             type="password"
-            placeholder="Paste your master key here…"
+            placeholder="Paste your UI access key here…"
             class="w-full rounded-lg border border-gray-300 px-4 py-2.5 text-sm text-gray-900 placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-mozilla-500 focus:border-transparent transition"
             autocomplete="off"
             spellcheck="false"
@@ -170,11 +170,11 @@
 
       try {
         const res = await fetch('/user/counts-by-service-type', {
-          headers: { 'master_key': `Bearer ${key}` },
+          headers: { 'mlpa_ui_access_key': `Bearer ${key}` },
         });
 
         if (res.status === 401) {
-          showError('Invalid master key — request was rejected (401 Unauthorized).');
+          showError('Invalid UI access key — request was rejected (401 Unauthorized).');
           return;
         }
         if (!res.ok) {
